+++ src/mykrobe/constants.py
@@ -3,3 +3,6 @@ import os
 DB_PREFIX = "atlas"
 STOP = "*"
 K = os.environ.get("KMER_SIZE", 21)
+ONT_E_RATE = 0.08
+ONT_PLOIDY = "haploid"
+ILLUMINA_E_RATE = 0.05